Core 2fa
This update has the core code to generate a 2fa secret for a user and present it as a in terminal qr code. It then has a function to take a user id and token and compute if its correct.

Next a clean up and implementation in to the login function is needed.
