+++ server/accounts/accountCreation.js
@@ -46,4 +46,4 @@ module.exports= {
     makeUser
 };
 
-//makeUser('lesasdt@tet.com','first','notfirst',"2022-12-20","nuts123").then((result) => console.log(result)).catch((err) => console.error(err));
+makeUser('lesasdt@tet.com','first','notfirst',"2022-12-20","nuts123").then((result) => console.log(result)).catch((err) => console.error(err));
